changing domain
I gave up on using a droplet to create a PHP 7 version of my blog because it would work out at about £2/month for not a lot of gain vs this free blog hosted on GitHub + the 2 year domain I got. https://tomwillgoto.space

DNS will have broken for up to 24 hours and I'm having to wait for the correct version of Ruby to compile on my 2010 MacBook so I can do some dev. This commit will break the site... for now.